Split large JSONata source file into smaller modules The JSONata source file is getting very long, and it would be better broken down into smaller files.  This PR: - Moves `/jsonata.js` to `/src` and splits off the numeric functions - Uses ES2015 module format so that   - Rollup can be used to join the files - A CommonJS version is genereted in `/lib`   - The tests are run against this version so that 100% code coverage can be achieved - A UMD version is produced where the original source file was
