+++ .eslintignore
@@ -1,5 +1,5 @@
 coverage/
 node_modules/
 doc/
-jsonata-es5.js
-*.min.js
+lib/
+/jsonata*.js
