+++ .npmignore
@@ -1,5 +1,6 @@
 *
-!*.js
+!jsonata*.js
 !LICENSE
 !package.json
 !README.md
+!lib/*
