+++ package.json
@@ -9,10 +9,11 @@
     "url": "https://github.com/jsonata-js/jsonata.git"
   },
   "scripts": {
-    "pretest": "npm run lint",
+    "pretest": "npm run lint && npm run rollup",
     "test": "node ./node_modules/istanbul/lib/cli.js cover --report cobertura --report html ./node_modules/mocha/bin/_mocha -- \"test/**/*.js\"",
     "posttest": "npm run check-coverage && npm run minify && npm run babel && npm run browserify && npm run minify-es5",
     "check-coverage": "istanbul check-coverage -statement 100 -branch 100 -function 100 -line 100",
+    "rollup": "rollup -c",
     "minify": "uglifyjs jsonata.js -o jsonata.min.js --compress --mangle",
     "lint": "eslint .",
     "postlint": "npm run doc",
@@ -48,6 +49,7 @@
     "mocha": "^3.4.2",
     "mocha-lcov-reporter": "^1.2.0",
     "request": "^2.81.0",
+    "rollup": "^0.43.0",
     "uglify-es": "^3.0.20"
   },
   "engines": {
