+++ test/.eslintrc
@@ -1,4 +1,8 @@
 {
+  "parserOptions": {
+    "sourceType": "script"
+  },
+
   // Rules ammendments
   "rules": {
     "strict": ["error", "global"],
